OrganisationController - implemented Create Medium endpoints

+++ src/Core/Givt.OnlineCheckout.Business/QR/Organisations/Texts/Create/CreateOrganisationTextsCommand.cs
@@ -5,7 +5,7 @@ namespace Givt.OnlineCheckout.Business.QR.Organisations.Texts.Create;
 
 public class CreateOrganisationTextsCommand : LocalisableTextsCore, IRequest<CreateOrganisationTextsResult>
 {
-    public int OrganisationId { get; set; }
+    public long OrganisationId { get; set; }
     public string LanguageId { get; set; }
 
 }
